finish up user system, fix some display problems only have to add per-user encryption + fix sendtoaddress now

+++ src/qt/transactiontablemodel.cpp
@@ -547,7 +547,7 @@ QVariant TransactionTableModel::data(const QModelIndex &index, int role) const
         {
             return COLOR_UNCONFIRMED;
         }
-        if(index.column() == Amount && (rec->credit+rec->debit) < 0)
+        if(index.column() == Amount && ((int64)rec->credit+rec->debit) < 0)
         {
             return COLOR_NEGATIVE;
         }
